==> scripts/01_Analisis_Exploratorio.Rmd <==
---
title: "01_Análisis_Exploratorio"
author: "Paula Sinisterra Sebastián"
date: "2023-04-04"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# Carga de librerías

library(readr) # Importar documentos
library(readxl) # Importar documentos
library(writexl) # Exportar documentos
library(dplyr) # Crear CODMUN
library(stringr) # Concatenar vectores
library(ggplot2) # Hacer gráficas
library(RColorBrewer) # Colores mapas

```

```{r}
# Carga de datos de TB
Datos_Global <- read_csv("../data/data_Paula_2014_2020_epi_clusters.csv")
Datos_Global <- as.data.frame(Datos_Global)

# Carga de datos de población CV
load("../data/CupoPobMunis.RData")
```


# Análisis exploratorio inicial

```{r}

# Se eliminan las 3 últimas columnas porque están repetidas
Datos_Global <- Datos_Global[,c(1:72)]

# Se transforma todo a factor
for (i in 2:72){
  Datos_Global[,i]<- as.factor(Datos_Global[,i])
}

# Se recuperan el formato de variables numéricas
Datos_Global$year_declaracion <- as.numeric(Datos_Global$year_declaracion)
Datos_Global$S_Declaracion <- as.numeric(Datos_Global$S_Declaracion)
Datos_Global$No_Edad <- as.numeric(Datos_Global$No_Edad)

# Se cambian los nombres que no son adecuados
colnames(Datos_Global)[1] <- "ID_Genom"
colnames(Datos_Global)[2] <- "Cluster_5snps"
colnames(Datos_Global)[3] <- "Cluster_10snps"
colnames(Datos_Global)[45] <- "resH.INH"
colnames(Datos_Global)[46] <- "resZ.PZA"
colnames(Datos_Global)[47] <- "resE.EMB"
colnames(Datos_Global)[48] <- "resR.RIF"
colnames(Datos_Global)[49] <- "resS.SM"
```

```{r}

# Exportamos .csv con los datos epidemiológicos que faltan
Faltantes_Epi <- Datos_Global[ is.na(Datos_Global$F_Declaracion) & !is.na(Datos_Global$ID_Genom),]
#write.csv2(Faltantes_Epi, "../results/Faltantes_Epi.csv", row.names = FALSE)

# Datos que sí tienen información epi de la encuesta 
Con_Epi <- Datos_Global[!is.na(Datos_Global$Sexo),]

# Se eliminan los individuos con NAs Municipio_Residencia
Con_Epi[is.na(Con_Epi$Municipio_Residencia),]

Datos_Bien <- Con_Epi[!is.na(Con_Epi$Municipio_Residencia),]
Datos_Bien <- Datos_Bien[Datos_Bien$Municipio_Residencia!="ALMANSA",]


# Hay 1407 datos que no tienen muestra genómica
Faltantes_Genom <- Datos_Global[!is.na(Datos_Global$F_Declaracion) & is.na(Datos_Global$ID_Genom),]
```


## Casos de TB por Año

```{r, include=TRUE, echo=FALSE}
Datos_Bien$year_declaracion <- as.factor(Datos_Bien$year_declaracion)
plot(Datos_Bien$year_declaracion, main="Casos de TB por año")
```

## Casos de TB por Sexo y casos por Año y Sexo

```{r}
# Número de casos notificados de TB por Sexo

Datos_Bien_Sexo <- table(Datos_Bien$Sexo)
Datos_Bien_Sexo <- as.data.frame(Datos_Bien_Sexo)
colnames(Datos_Bien_Sexo)[1] <- "Sexo"
colnames(Datos_Bien_Sexo)[2] <- "Casos_TB"
p1 <- ggplot(Datos_Bien_Sexo, aes(x = Sexo, y = Casos_TB)) +
  geom_bar(stat = "identity") +
  labs(title='Número de casos notificados de TB por Sexo') +
  theme(plot.title=element_text())
p1
```

```{r}
# Número de casos notificados de TB por año y sexo

Sexo_Año <- table(Datos_Bien$Sexo, Datos_Bien$year_declaracion)
colnames(Sexo_Año) <- c(2014:2020)

Sexo_Año <- as.data.frame(Sexo_Año)
colnames(Sexo_Año)[1] <- "Sexo"
colnames(Sexo_Año)[2] <- "Año"
colnames(Sexo_Año)[3] <- "Casos_TB"

p2 <- ggplot(data=Sexo_Año, aes(x=Año, y=Casos_TB, fill=Sexo)) +
  geom_bar(stat="identity", position=position_dodge())+
  geom_text(aes(label=Casos_TB), vjust=1.6, color="white",
            position = position_dodge(0.9), size=3.5)+
  scale_fill_brewer(palette="Paired")+
  labs(title='Número de casos notificados de TB por año y sexo')+
  theme_minimal()+
  theme(plot.caption = element_text(hjust = 0, size=10))

p2
```

## Mediana de edad en Hombres y mujeres

```{r, include=TRUE, echo=FALSE, fig.align='center', out.width="90%"}

# Mediana de edad en Hombres y mujeres
edad_h <- Datos_Bien[Datos_Bien$Sexo=="HOMBRE",]
edad_m <- Datos_Bien[Datos_Bien$Sexo=="MUJER",]

median(edad_h$No_Edad)
median(edad_m$No_Edad)

# Test de Wilcoxon para diferencia entre medianas
wilcox.test(edad_h$No_Edad, edad_m$No_Edad)

# Test-t para diferencia entre medias
t.test(edad_h$No_Edad, edad_m$No_Edad)
  
hist(edad_h$No_Edad)
hist(edad_m$No_Edad)
hist(Datos_Bien$No_Edad)
```

## Distribución de casos de TB por edades
```{r, include=FALSE}
Datos_Bien$No_Edad  <- Datos_Bien$No_Edad - 1
Datos_Bien$Grupo_Edad <- cut(as.numeric(Datos_Bien$No_Edad), 
                             breaks = c(0,1,5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,100), 
                             include.lowest = TRUE, right=FALSE)

df_Edad <- as.data.frame(table(Datos_Bien$Grupo_Edad))
colnames(df_Edad)[1] <- "Grupo_Edad"
colnames(df_Edad)[2] <- "Casos_TB"
p3 <- ggplot(df_Edad, aes(x = Grupo_Edad, y = Casos_TB)) +
  geom_bar(stat = "identity") +
  labs(title='Número de casos notificados de TB por grupo de edad') +
  theme(plot.title=element_text(), axis.text = element_text(angle = 45))

p3

Sexo_Edad <- table(Datos_Bien$Sexo, Datos_Bien$Grupo_Edad)
Sexo_Edad <- as.data.frame(Sexo_Edad)
colnames(Sexo_Edad)[1] <- "Género"
colnames(Sexo_Edad)[2] <- "Grupo_Edad"
colnames(Sexo_Edad)[3] <- "Casos_TB"


p4 <- ggplot(data=Sexo_Edad, aes(x=Grupo_Edad, y=Casos_TB, fill=Género)) +
  geom_bar(stat="identity", position=position_dodge())+
  scale_fill_brewer(palette="Paired")+
  labs(title='Número de casos notificados de TB por grupo de edad y género') +
  theme_minimal()+
  theme(plot.caption = element_text(hjust = 0, vjust=0, size=10), axis.text = element_text(angle = 45))

p4

```

```{r}

# Pirámide de población de los casos de TB notificados en la CV
Sexo_Edad <- as.data.frame(Sexo_Edad)
colnames(Sexo_Edad)[1] <- "Sexo"
colnames(Sexo_Edad)[2] <- "Edad"
colnames(Sexo_Edad)[3] <- "Poblacion"
str(Sexo_Edad)

Sexo_Edad <- Sexo_Edad %>%
  mutate(
    Poblacion = ifelse(
      Sexo == "HOMBRE",
      -Poblacion, Poblacion)
    )

```

```{r}


p5 <-  ggplot(Sexo_Edad,
    aes(x = Edad,
        y = Poblacion, fill = Sexo)
    ) +
  # agregar el argumento color = "white", si gustan
    geom_col(position = "stack", alpha = 0.6) + 
    coord_flip() +
  # colores que me gustan
    scale_fill_manual(values = c("midnightblue", "darkred")) +
  # tema minimalista
    theme_minimal() +
  # leyenda para el fondo
    theme(
      legend.position = "bottom",
      plot.caption = element_text(hjust = 0)) +
  # etiquetas en positivo
    scale_y_continuous(labels = abs) +
    labs(
      y = "Población afectada por TB en la CV (Nº de individuos)",
      x = "Quinquenios de edad",
      title = "Pirámide de población de TB", 
      subtitle = "Edad en grupos quinquenales",
       tag = "A")

p5
```
```{r}

# Pirámide de población de la CV
CV_Edad_Sexo <- apply(cubopob[,,,15:21],2:3,sum)

Sexo1 <- CV_Edad_Sexo[,1]
Sexo1 <- as.data.frame(Sexo1)
Sexo1$Sexo <- c("HOMBRE")
Sexo1$Poblacion <- rownames(Sexo1)
rownames(Sexo1) <- NULL
colnames(Sexo1)[1] <- "Poblacion"
colnames(Sexo1)[3] <- "Grupo.Edad"

Sexo6 <- CV_Edad_Sexo[,2]
Sexo6 <- as.data.frame(Sexo6)
Sexo6$Sexo <- c("MUJER")
Sexo6$Poblacion <- rownames(Sexo6)
rownames(Sexo6) <- NULL
colnames(Sexo6)[1] <- "Poblacion"
colnames(Sexo6)[3] <- "Grupo.Edad"

Sexo7 <- CV_Edad_Sexo[,3]
Sexo7 <- as.data.frame(Sexo7)
Sexo7$Sexo <- c("7")
Sexo7$Poblacion <- rownames(Sexo7)

CV_Sexo_Edad <- rbind(Sexo1,Sexo6)

```
```{r}
CV_Sexo_Edad <- CV_Sexo_Edad %>%
  mutate(
    Poblacion = ifelse(
      Sexo == "HOMBRE",
      -Poblacion, Poblacion)
    )

CV_Sexo_Edad$Grupo.Edad[2] <- "01-04"
CV_Sexo_Edad$Grupo.Edad[3] <- "05-09"
CV_Sexo_Edad$Grupo.Edad[21] <- "01-04"
CV_Sexo_Edad$Grupo.Edad[22] <- "05-09"

```
```{r}

p6 <- ggplot(CV_Sexo_Edad,
    aes(x = Grupo.Edad,
        y = Poblacion, fill = Sexo)
    ) +
  # agregar el argumento color = "white", si gustan
    geom_col(position = "stack", alpha = 0.6) + 
    coord_flip() +
  # colores que me gustan
    scale_fill_manual(values = c("midnightblue", "darkred")) +
  # tema minimalista
    theme_minimal() +
  # leyenda para el fondo
    theme(
      legend.position = "bottom",
      plot.caption = element_text(hjust = 0)) +
  # etiquetas en positivo
    scale_y_continuous(labels = abs) +
    labs(
      y = " Población de riesgo en la CV (Nº de individuos)",
      x = "Quinquenios de edad",
      title = "Pirámide de población de riesgo CV", 
      subtitle = "Edad en grupos quinquenales",
       tag = "B")

p6
```

```{r}

# Pirámide de tasas de TB en la CV

rm(Sexo_Edad)
Sexo_Edad <- table(Datos_Bien$Sexo, Datos_Bien$Grupo_Edad)
CV_Sexo_Edad <- apply(cubopob,c(3,2),sum)[c(1,2),]

Tasas_Edad <- Sexo_Edad/CV_Sexo_Edad 
Tasas_Edad <- Tasas_Edad*100000

Tasas_Edad <- as.data.frame(Tasas_Edad)
colnames(Tasas_Edad)[1] <- "Sexo"
colnames(Tasas_Edad)[2] <- "Edad"
colnames(Tasas_Edad)[3] <- "Tasa"

```

```{r}
Tasas_Edad <- Tasas_Edad %>%
  mutate(
    Tasa = ifelse(
      Sexo == "HOMBRE",
      -Tasa, Tasa)
    )
```
```{r}

labs(title='Número de casos notificados de TB por grupo de edad y género', tag = "B", caption = "Figura 2. A) Representación gráfica del número de casos notificados de TB entre 2014 y 2020 en la CV
según el grupo de edad. B) Representación gráfica del número de casos notificados de TB entre 2014 
y 2020 en la CV según el grupo de edad y el año.") +
  theme_minimal()+
  theme(plot.caption = element_text(hjust = 0, vjust=0, size=10))

p7 <- ggplot(Tasas_Edad,
    aes(x = Edad,
        y = Tasa, fill = Sexo)
    ) +
  # agregar el argumento color = "white", si gustan
    geom_col(position = "stack", alpha = 0.6) + 
    coord_flip() +
  # colores que me gustan
    scale_fill_manual(values = c("midnightblue", "darkred")) +
  # tema minimalista
    theme_minimal() +
  # leyenda para el fondo
    theme(
      legend.position = "bottom",
      plot.caption = element_text(hjust = 0)) +
  # etiquetas en positivo
    scale_y_continuous(labels = abs) +
    labs(
      y = " Tasa de TB por 100000 habitantes en la CV",
      x = "Quinquenios de edad",
      title = "Pirámide de TASAS de TB por 100 000 habitantes", 
      subtitle = "Edad en grupos quinquenales",
      tag = "C")

p7
```
## Casos de TB declarados por año según tipo de caso (extranjero/español)
```{r}
Datos_Bien_Extranjero <- table(Datos_Bien$Extranjero)
Datos_Bien_Extranjero <- as.data.frame(Datos_Bien_Extranjero)
colnames(Datos_Bien_Extranjero)[1] <- "Extranjero"
colnames(Datos_Bien_Extranjero)[2] <- "Casos_TB"

p8 <- ggplot(Datos_Bien_Extranjero, aes(x = Extranjero, y = Casos_TB)) +
  geom_bar(stat = "identity") +
  labs(title='Número de casos de TB en individuos extranjeros y españoles') +
  theme(plot.title=element_text())


Extranjero_Año <- table(Datos_Bien$Extranjero, Datos_Bien$year_declaracion)
colnames(Extranjero_Año) <- c(2014:2020)

Extranjero_Año <- as.data.frame(Extranjero_Año)
colnames(Extranjero_Año)[1] <- "Extranjero"
colnames(Extranjero_Año)[2] <- "Año"
colnames(Extranjero_Año)[3] <- "Casos_TB"

# Total de casos declarados por año según tipo de caso (extranjero/español) y año

p9 <- ggplot(data=Extranjero_Año, aes(x=Año, y=Casos_TB, fill=Extranjero)) +
  geom_bar(stat="identity", position=position_dodge())+
  geom_text(aes(label=Casos_TB), vjust=1.6, color="white",
            position = position_dodge(0.9), size=3.5)+
  scale_fill_brewer(palette="Paired")+
  labs(title='Número de casos de TB en individuos extranjeros y españoles por año ')+
  theme_minimal()+
  theme(plot.caption = element_text(hjust = 0))

p8
p9
```
```{r}

# Tasas de TB según tipo de caso (extranjero/español)

Year <- c(2014:2020)
CV <- c(5004844, 4980689, 4959968, 4941509,4963703, 5003769, 5057353)
P_Ext <- c(739630, 700211, 672379, 642380, 664921,703413,752131)
P_Nac <- CV - P_Ext

Extranjero <- c("NO", "NO", "NO", "NO")
CV_Year <- data.frame("Year"=Year,"Nac"=P_Nac)
CV_Year <- as.matrix(CV_Year)

Extranjero <- c("SI", "SI", "SI", "SI")
Ext_Year <- data.frame("Year"=Year,"CV"=P_Ext)
Ext_Year <- as.matrix(Ext_Year)

CV_Year <- t(CV_Year)
Ext_Year <- t(Ext_Year)

Extranjeros_CV  <- rbind(CV_Year,Ext_Year[2,])

colnames(Extranjeros_CV) <- c(Extranjeros_CV[1,])
Extranjeros_CV <- Extranjeros_CV[c(2,3),]
rownames(Extranjeros_CV )[2] <- "Extr"


## Casos de TB 

Extranjeros_TB <- table(Datos_Bien$Extranjero, Datos_Bien$year_declaracion)
Tasas_Ext <- Extranjeros_TB/Extranjeros_CV *100000
colnames(Tasas_Ext) <- c(2014:2020)

Tasas_Ext<- round(Tasas_Ext,2)
Tasas_Ext <- as.data.frame(Tasas_Ext)

colnames(Tasas_Ext)[1] <- "Extranjero"
colnames(Tasas_Ext)[2] <- "Año"
colnames(Tasas_Ext)[3] <- "Tasa_TB"

ggplot(data=Tasas_Ext, aes(x=Año, y=Tasa_TB, fill=Extranjero)) +
  geom_bar(stat="identity", position=position_dodge())+
  geom_text(aes(label=Tasa_TB), vjust=1.6, color="white",
            position = position_dodge(0.9), size=3.5)+
  scale_fill_brewer(palette="Paired")+
  labs(title='Tasa de casos de TB por 100000 habitantes en locales y extranjeros', )+
  theme_minimal()+
  theme(plot.caption = element_text(hjust = 0))
```
## Relación TB y VIH
```{r}
# Tasas VIH CV
VIH_CV <- (350+317+317+382+348 + 485 + 364)/7
CV <- (5004844+ 4980689+ 4959968+ 4941509+4963703+ 5003769+ 5057353)/7

VIH_CV/CV

# Coinfección con VIH
table(Datos_Bien$VIH)
sum(is.na(Datos_Bien$VIH))
```

## Casos de TB con información genómica asociada

```{r, results='hide'}
# Sacamos un data.frame con los datos de TB que tienen información
# genómica asociada
Datos_Genom <- Datos_Bien[ !is.na(Datos_Bien$ID_Genom),]

# Se crea una variable Bernoulli que nos diga si están o no en transmisión
Datos_Genom$Ber <- Datos_Genom$Cluster_5snps

Datos_Genom$Ber <- as.character(Datos_Genom$Ber)

Datos_Genom$Ber[Datos_Genom$Ber=="unique"] <- "NO"
Datos_Genom$Ber[Datos_Genom$Ber!="NO"] <- "SI"

Datos_Genom$Ber <- as.factor(Datos_Genom$Ber)
plot(Datos_Genom$Ber,  main="Casos en Transmisión Genómica")

colSums(is.na(Datos_Genom)) #MunRes 0 NAs

```
