==> scripts/02_Preparacion_Riesgos.Rmd <==
---
title: "02_Preparación_Riesgos"
author: "Paula Sinisterra Sebastián"
date: "2023-06-05"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
# Carga de librerías

library(readr) # Importar documentos
library(readxl) # Importar documentos
library(writexl) # Exportar documentos
library(dplyr) # Crear CODMUN
library(stringr) # Concatenar vectores
```
```{r}
# Carga de datos de TB
Datos_Global <- read_csv("../data/data_Paula_2014_2020_epi_clusters.csv")
Datos_Global <- as.data.frame(Datos_Global)

# Carga de datos de población CV
load("../data/CupoPobMunis.RData")
```

```{r}

# Se eliminan las 3 últimas columnas porque están repetidas
Datos_Global <- Datos_Global[,c(1:72)]

# Se transforma todo a factor
for (i in 2:72){
  Datos_Global[,i]<- as.factor(Datos_Global[,i])
}

# Se recuperan el formato de variables numéricas
Datos_Global$year_declaracion <- as.numeric(Datos_Global$year_declaracion)
Datos_Global$S_Declaracion <- as.numeric(Datos_Global$S_Declaracion)
Datos_Global$No_Edad <- as.numeric(Datos_Global$No_Edad)

# Se cambian los nombres que no son adecuados
colnames(Datos_Global)[1] <- "ID_Genom"
colnames(Datos_Global)[2] <- "Cluster_5snps"
colnames(Datos_Global)[3] <- "Cluster_10snps"
colnames(Datos_Global)[45] <- "resH.INH"
colnames(Datos_Global)[46] <- "resZ.PZA"
colnames(Datos_Global)[47] <- "resE.EMB"
colnames(Datos_Global)[48] <- "resR.RIF"
colnames(Datos_Global)[49] <- "resS.SM"
```

```{r}

# Se exporta un .csv con los datos epidemiológicos que faltan
Faltantes_Epi <- Datos_Global[ is.na(Datos_Global$F_Declaracion) & !is.na(Datos_Global$ID_Genom),]
#write.csv2(Faltantes_Epi, "../results/Faltantes_Epi.csv", row.names = FALSE)

# Datos que sí tienen información epi de la encuesta 
Con_Epi <- Datos_Global[!is.na(Datos_Global$Sexo),]

# Se eliminan los individuos con NAs Municipio_Residencia
Con_Epi[is.na(Con_Epi$Municipio_Residencia),]

Datos_Bien <- Con_Epi[!is.na(Con_Epi$Municipio_Residencia),]
Datos_Bien <- Datos_Bien[Datos_Bien$Municipio_Residencia!="ALMANSA",]

# Hay 1407 datos que no tienen muestra genómica
Faltantes_Genom <- Datos_Global[!is.na(Datos_Global$F_Declaracion) & is.na(Datos_Global$ID_Genom),]
```
```{r}
# Se crea la variable grupo de edad
Datos_Bien$No_Edad  <- Datos_Bien$No_Edad - 1 #Para que coincida con edad
Datos_Bien$Grupo_Edad <- cut(as.numeric(Datos_Bien$No_Edad), 
                             breaks = c(0,1,5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,100), 
                             include.lowest = TRUE, right=FALSE)
```

# Preparación de los datos de TB para aplicar los modelos de riesgo

```{r, results='hide'}
# Sacamos un data.frame con los datos de TB que tienen información
# genómica asociada
Datos_Genom <- Datos_Bien[ !is.na(Datos_Bien$ID_Genom),]

# Se crea una variable Bernoulli que nos diga si están o no en transmisión
Datos_Genom$Ber <- Datos_Genom$Cluster_5snps

Datos_Genom$Ber <- as.character(Datos_Genom$Ber)

Datos_Genom$Ber[Datos_Genom$Ber=="unique"] <- "NO"
Datos_Genom$Ber[Datos_Genom$Ber!="NO"] <- "SI"

Datos_Genom$Ber <- as.factor(Datos_Genom$Ber)
plot(Datos_Genom$Ber,  main="Casos en Transmisión Genómica")

colSums(is.na(Datos_Genom)) #MunRes 0 NAs
```


## Obtención de vectores de casos de TB observados y esperados

Para aplicar cualquier modelo de riesgo de TB se debe contar con:

* Vector con casos de TB por municipio 
* Vector con casos esperados de TB por municipio.

A su vez, para calcular los casos esperados por municipio, se utiliza la población de la CV en grupos de sexo y edad y las tasas de TB de la CV asociadas a casa uno de los grupos de sexo y edad. Para obtenerlas se realizan los cáculos siguientes:


### Población por municipio en la CV

```{r}
# Se carga el array del que disponemos con la info de la CV

load("../data/CupoPobMunis.RData")

class(cubopob)
dim(cubopob)

```

### Obtención de matrices POPT, POPT.1 y POPT.2

```{r}
# Se crea la matriz POPT (que contiene en cada casilla el número de personas año desde 2014 a 2020 de ambos sexos)
POP <- cubopob[,,3,15:21]
POPT <- apply(POP,1:2,sum)

# Se crea una matriz con las personas·año de sexo 1 (hombres) 
# en cada municipio divididas por edades.

POP1 <- cubopob[,,1,15:21]
POPT.1 <- apply(POP1,1:2,sum) 

# Se crea una matriz con las personas·año de sexo 6 (mujeres) 
# en cada municipio divididas por edades.
POP2 <- cubopob[,,2,15:21]
POPT.2 <- apply(POP2,1:2,sum) 

# Se crea un vector con todos los nombres de los municipios de la CV
Todo_Ine <- as.data.frame(rownames(POP))
```


### Creación de la variable CODMUN en la base de datos principal

```{r}
# Lista con los códigos INE de cada municipio ordenados alfabéticamente
CodINE <- read_excel("../data/20codmun.xlsx")

# Creamos la variable del código INE
CodINE_CV <- filter(CodINE, CODAUTO=="10")
CodINE_CV$INE <- str_c(CodINE_CV$CPRO,CodINE_CV$CMUN)

# Ordenamos los municipios alfabéticamente
CodINE_CV_Alf <- CodINE_CV[order(CodINE_CV$NOMBRE),c(5:6)]
colnames(CodINE_CV_Alf)[1] <- "Municipio_Residencia"
CodINE_CV_Alf$Municipio_Residencia <- as.factor(CodINE_CV_Alf$Municipio_Residencia)

# Exportamos la lista
#write_xlsx(CodINE_CV_Alf,"../results/CodINE.CV.Alf.xlsx" )
```

Para poder poner los códigos cruzando las tablas, se debe asegurar que los nombres de los niveles de la variable del municipio de residencia son iguales en ambas bases de datos. Para ello, en la base de datos 20codmun.xlsx se han:

* Cambiado los nombres a mayúsculas
* Eliminado los acentos
* Sustituido los Artículos por (Artículos)

Además, a continuación, se realizan algunos cambios en los nombres de la base de Datos_Bien para terminar de completar. 

```{r}

Datos_Bien$Municipio_Residencia <- as.character(Datos_Bien$Municipio_Residencia)

Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="ADALIA"] <- "ALDAIA"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="ALQUERIA DE LA COMTESSA ("] <- "ALQUERIA DE LA COMTESSA (L')"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="ALMAZORA/ALMASSORA"] <- "ALMASSORA"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="ALQUERIAS DEL NIÃ‘O PERDID"] <- "ALQUERIES, LES/ALQUERIAS DEL NIÑO PERDIDO"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="BUÃ‘OL"] <- "BUÑOL"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="BURRIANA"] <- "BORRIANA/BURRIANA"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="CABANES (CASTELLoN)"] <- "CABANES"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="EL REAL DE GANDIA"] <- "REAL DE GANDIA (EL)"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="FONDO DE LES NEUS (EL)/HO"] <- "FONDO DE LES NEUS (EL)/HONDON DE LAS NIEVES"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="PEÃ‘ISCOLA"] <- "PENISCOLA/PEÑISCOLA"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="PUÃ‡OL"] <- "PUÇOL"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="PUIG DE SANTA MARIA Eel"] <- "PUIG DE SANTA MARIA (EL)"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="PUIG DE SANTA MARIA EL"] <- "PUIG DE SANTA MARIA (EL)"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="SAN VICENTE DEL RASPEIG"] <- "SAN VICENTE DEL RASPEIG/SANT VICENT DEL RASPEIG"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="TORREMANZANAS/TORRE DE LE"] <- "TORREMANZANAS/TORRE DE LES MAÇANES (LA)"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="TORRENT (VALENCIA)"] <- "TORRENT"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="VILLAFRANCA DEL CID/VILAF"] <- "VILAFRANCA/VILLAFRANCA DEL CID"
Datos_Bien$Municipio_Residencia[Datos_Bien$Municipio_Residencia=="VILLAJOYOSA/VILA JOIOSA ("] <- "VILLAJOYOSA/VILA JOIOSA (LA)"

Datos_Bien$Municipio_Residencia <- as.factor(Datos_Bien$Municipio_Residencia)
Datos_Bien$Municipio_Residencia  <- droplevels(Datos_Bien$Municipio_Residencia )

``` 

```{r}

# Una vez unificados los nombres se cruzan las bases de datos
Datos_Codigos <- left_join(Datos_Bien,CodINE_CV_Alf, by="Municipio_Residencia")

Datos_Codigos[Datos_Codigos$Municipio_Residencia=="ALMANSA",]
Datos_Codigos[is.na(Datos_Codigos$INE),]

# Eliminamos al individuo que reside en ALMANSA, ya que no es CV
Datos_Codigos <- Datos_Codigos[!is.na(Datos_Codigos$INE),]
```
```{r}
# Imprimimos base de datos sin NAs y con Códigos
# write.csv(Datos_Codigos, "../data/Datos_Codigos.csv")
```


### Obtención de casos observados por municipio (O_i)

```{r}

# Número de nuevos casos observados en  en cada municipio (O_i) 

colSums(is.na(Datos_Codigos))# No hay datos NA
Obs_Muni <- as.data.frame(table(Datos_Codigos$INE))

colnames(Todo_Ine)[1] <- "Var1"

Obs_Muni <- left_join(x = Todo_Ine, y = Obs_Muni,by="Var1") 
Obs_Muni [is.na(Obs_Muni)] <- 0
```

### Cálculo de Tasas de TB por edad y sexo en la CV

```{r}
 
# Matriz con el número de casos observados para cada edad y cada sexo.
 
Obs_Age_Sex <- as.matrix(table(Datos_Codigos$Sexo, Datos_Codigos$Grupo_Edad)) 

# Tasas para cada edad y para cada sexo

# Suma de todas las personas del sexo 1 de la CV en cada rango de edad
Age_Sex_1 <- apply(POPT.1,2,sum)
# Suma de todas las personas del sexo 6 de la CV en cada rango de edad
Age_Sex_2 <- apply(POPT.2,2,sum)

CV_Age_Sex <- as.data.frame(Age_Sex_1)
CV_Age_Sex$Age_Sex_2 <- Age_Sex_2

CV_Age_Sex_t <- as.matrix(t(CV_Age_Sex))

# Se dividen los observados de TB entre la población total para sacar la tasa
Rates_AS <- Obs_Age_Sex/CV_Age_Sex_t 
```

### Cálculo de los casos esperados en cada municipio

```{r}

# Primero se calculan los esperados para cada uno de los sexos
# Y luego se suman ambos resultados para obtener los esperados totales

# Casos esperados para cada rango de edad en Sexo1
POPT.1 <- as.matrix(POPT.1)
Rates_1 <- as.matrix(t(Rates_AS))[,1] 
Exp_Age_1 <- as.vector(POPT.1 %*% Rates_1) 

# Casos esperados para cada rango de edad en Sexo6
POPT.2 <- as.matrix(POPT.2)
Rates_2 <- as.matrix(t(Rates_AS))[,2]
Exp_Age_2 <- as.vector(POPT.2 %*% Rates_2) 

# Se suman los casos de ambos sexos
Exp_Age_Sex <- data.frame("1"=Exp_Age_1, "2"=Exp_Age_2)
Exp_Age_Sex <- as.matrix(Exp_Age_Sex)

Exp_Muni_AS <- apply(Exp_Age_Sex,1,sum)
Exp_Muni_AS <- as.data.frame(Exp_Muni_AS)

# Se multiplican las tasas-año por las personas año en cada municipio
Exp_Muni_AS <- as.data.frame(Exp_Muni_AS) 
POPT.1 <- as.data.frame(POPT.1)
Exp_Muni_AS$Var1<- rownames(POPT.1)
colnames(Exp_Muni_AS)[1] <- "Freq"

Exp_Muni_AS <- Exp_Muni_AS [ , rev(order(c(names(Exp_Muni_AS))))]
```


```{r}
# Se exportan los vectores de observados y esperados

#write_xlsx(Obs_Muni,"../results/Obs_Muni.xlsx" )
#write_xlsx(Exp_Muni_AS,"../results/Exp_Muni_AS.xlsx" )
```
