==> scripts/05_Analisis_Comparativo.Rmd <==
---
title: "05_Analisis_Complementarios"
author: "Paula Sinisterra Sebastián"
date: "2023-04-13"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, include = FALSE)
```


```{r, include=FALSE}
library(readr) # Importar documentos
library(readxl) # Importar documentos
library(knitr)
library(kableExtra) # Creación de tablas
library(dplyr)
```

# Análisis comparativo entre la muestra de casos de TB con información genómica y la muestra de casos de TB sin información genómica

## Carga y preparación de los datos

```{r}
# Base de datos arreglada
Datos_Codigos <- read_csv("../data/Datos_Codigos.csv")
Datos_Codigos <- as.data.frame(Datos_Codigos[,2:75])

# Hay 1400 datos que no tienen muestra genómica
Faltantes_Genom <- Datos_Codigos[!is.na(Datos_Codigos$F_Declaracion) & is.na(Datos_Codigos$ID_Genom),]

# Sacamos un data.frame con los datos genómicos
Datos_Genom <- Datos_Codigos[ !is.na(Datos_Codigos$ID_Genom),]

Datos_Genom$Ber <- Datos_Genom$Cluster_5snps
Datos_Genom$Ber[Datos_Genom$Ber=="unique"] <- "NO"
Datos_Genom$Ber[Datos_Genom$Ber!="NO"] <- "SI"
```

```{r}
for (i in 2:74){
  Datos_Codigos[,i]<- as.factor(Datos_Codigos[,i])
}

for (i in 2:74){
  Faltantes_Genom[,i]<- as.factor(Faltantes_Genom[,i])
}

for (i in 2:75){
  Datos_Genom[,i]<- as.factor(Datos_Genom[,i])
}
```

## Análisis del número de casos de TB sin información genómica asociada por provincia y departamento de salud

### Por provincia

```{r}
Datos_Genom[is.na(Datos_Genom$Provincia_Declaracion),]
table(Datos_Genom$Provincia_Declaracion)
```

```{r}
# Cargamos vector de casos observados
Obs_Muni_Genom <- read_excel("../results/Obs_Muni_Genom.xlsx")

# Cargamos los datos de población
load("../data/CupoPobMunis.RData")
POP <- cubopob[,,3,]
Todo_Ine <- as.data.frame(rownames(POP))
colnames(Todo_Ine)[1] <- "Var1"
```

```{r}
# Se calculan los casos de TB observados sin info genómica
# en cada provincia

Obs_Muni_Faltantes<- as.data.frame(table(Faltantes_Genom$INE))
Obs_Muni_Faltantes <- left_join(x = Todo_Ine, y = Obs_Muni_Faltantes, by="Var1") 
Obs_Muni_Faltantes [is.na(Obs_Muni_Faltantes)] <- 0

sum(Obs_Muni_Faltantes[1:141,]$Freq)
sum(Obs_Muni_Faltantes[142:276,]$Freq)
sum(Obs_Muni_Faltantes[277:542,]$Freq)

# Dividiendo los casos de TB observados sin info genómica entre los casos totales
# se calculan las tasas de casos de TB sin info genómica

Tasa_F_Alc <- sum(Obs_Muni_Faltantes[1:141,]$Freq)/(sum(Obs_Muni_Faltantes[1:141,]$Freq)+sum(Obs_Muni_Genom[1:141,]$Freq))
Tasa_F_Cast <- sum(Obs_Muni_Faltantes[142:276,]$Freq)/(sum(Obs_Muni_Faltantes[142:276,]$Freq)+sum(Obs_Muni_Genom[142:276,]$Freq))
Tasa_F_Vlc <- sum(Obs_Muni_Faltantes[277:542,]$Freq)/(sum(Obs_Muni_Faltantes[277:542,]$Freq)+sum(Obs_Muni_Genom[277:542,]$Freq))

Tasas_Provincia <- data.frame(Provincia=c("Alicante", "Castellon", "Valencia"))
Tasas_Provincia$Tasa <- c(Tasa_F_Alc, Tasa_F_Cast, Tasa_F_Vlc)
```

```{r, include=TRUE, echo=FALSE}
kable(Tasas_Provincia, "latex", 
      caption = "\\label{tabla01} Tasas de casos de TB sin información genómica asociada por provincia",
      booktabs = T, 
      digits=c(0,4), 
      label="", 
      align = c("l","c")) %>%
column_spec(3, width = "1cm") 
```

```{r, include=TRUE}
# test de proporciones entre las tasas de casos de TB sin info. genómica
# entre las 3 provincias

x1 <- sum(Obs_Muni_Faltantes[1:141,]$Freq)
x2 <- sum(Obs_Muni_Faltantes[142:276,]$Freq)
x3 <- sum(Obs_Muni_Faltantes[277:542,]$Freq)

n1 <- sum(Obs_Muni_Faltantes[1:141,]$Freq)+sum(Obs_Muni_Genom[1:141,]$Freq)
n2 <- sum(Obs_Muni_Faltantes[142:276,]$Freq)+sum(Obs_Muni_Genom[142:276,]$Freq)
n3 <- sum(Obs_Muni_Faltantes[277:542,]$Freq)+sum(Obs_Muni_Genom[277:542,]$Freq)

prop.test(x=c(x1,x2), n=c(n1,n2),alternative="two.side", conf.level=0.99, correct=TRUE )
prop.test(x=c(x1,x3), n=c(n1,n3),alternative="two.side", conf.level=0.99, correct=TRUE )
prop.test(x=c(x2,x3), n=c(n2,n3),alternative="two.side", conf.level=0.99, correct=TRUE )
```


### Por departamento de salud

*Nota*: se utiliza el departamento de riesgo como variable para calculas las tasas de datos de TB sin información genómica porque esta variable en lugar de representar el departamento de salud donde se declara un caso, representa el departamento al que corresponde por residencia dicho caso. Por eso se considera que esta variable representa un poco mejor la realidad.

```{r}
Faltantes_Genom_DepR <- table(Faltantes_Genom$Dpto_Riesgo)
Faltantes_Genom_DepR <- as.data.frame(Faltantes_Genom_DepR)

Datos_Codigos_DepR <- table(Datos_Codigos$Dpto_Riesgo)
Datos_Codigos_DepR <- as.data.frame(Datos_Codigos_DepR)

Faltantes_Mun_DepR <- table(Faltantes_Genom$Dpto_Riesgo)/table(Datos_Codigos$Dpto_Riesgo)
Faltantes_Mun_DepR <- as.data.frame(Faltantes_Mun_DepR) 

Faltantes_MunR <- data.frame("Departamento"= Faltantes_Genom_DepR$Var1 ,
                            "Faltantes_Genom"=Faltantes_Genom_DepR$Freq, 
                            "Datos_Global"= Datos_Codigos_DepR$Freq, 
                            "Tasa Faltantes"=Faltantes_Mun_DepR$Freq)

Faltantes_MunR_Orden <- Faltantes_MunR[order(Faltantes_MunR$Tasa.Faltantes),]
```

```{r, include=TRUE, echo=FALSE}
kable(Faltantes_MunR_Orden, "latex", 
      caption = "\\label{tabla01} Tasas de datos Faltantes por Departamento de salud",
      booktabs = T, 
      digits=c(0,4,4,4), 
      label="", 
      align = c("l","c","c","c")) %>%
column_spec(1, width = "1cm") %>%
   kable_styling(full_width = FALSE, position = "center", font_size = 8)

```


## Análisis comparativo de diferentes factores de riesgo de la TB 

En segundo lugar, se realizó un análisis para ver comprobar si existían diferencias significativas en la proporción de otros factores de riesgo para la TB entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada. En todos los casos se utilizó el test de proporciones. Los factores que se han analizado, por disponibilidad y por suponer factores de riesgo para la TB son: el sexo, concretamente la proporción de hombres de las muestras, la condición de extranjero, haber estado ingresado en un hospital, ser personal sanitario, presentar alcoholismo, diabetes, haber consumido drogas por vía intravenosa (UDPV), presentar VIH, estar en situación de marginación social, haber pasado por prisión y estar en una residencia de ancianos. 

### A nivel global (CV)


```{r}
# Test de proporciones para la variable Sexo

Datos_Genom <- Datos_Genom[,1:74]
Faltantes_Genom <- Faltantes_Genom[,1:74]
Datos_Genom$Poblacion <- rep(1,1651)
Faltantes_Genom$Poblacion <- rep (2, 1400)

# Base de datos para trabajar
Analisis_Poblacion <- rbind(Datos_Genom,Faltantes_Genom)
```

```{r}

# Función para variables de 2 categorías (Sí/No)

## Sexo

test_prop_2 <- function (var1,var2){
  frec <- table(var1,var2)
  tot <- apply(frec,1,sum)
  return(list(frec,tot,prop.test(x=c(frec[,1]), n=c(tot), alternative="two.side", conf.level=0.99, correct=TRUE)))
}

Sexo_Global <-test_prop_2(Analisis_Poblacion$Poblacion, Analisis_Poblacion[,13]) #Comprobada

# Se modifica la función para que tome el valor de extranjero y no de "no extranjero"

test_prop_2 <- function (var1,var2){
  frec <- table(var1,var2)
  tot <- apply(frec,1,sum)
  return(list(frec,tot,prop.test(x=c(frec[,2]), n=c(tot), alternative="two.side", conf.level=0.99, correct=TRUE)))
}

objeto <- c()
for (i in c(14,50,63)){
  objeto[i] <- list(test_prop_2(Analisis_Poblacion$Poblacion, Analisis_Poblacion[,i]))
  cat ("Tabla y test de proporciones de la variable",names(Analisis_Poblacion)[i], "\n")
  print(objeto[i])
}

```

```{r}
# Variables de 3 categorías (Sí/No/Desc)
# Se elimina la categoría Desc porque equivale a NAs

x1 <- table(Analisis_Poblacion$Poblacion, Analisis_Poblacion$Diabetes)[,2:3] #Aquí se corrige el nivel desconocido
x1 <- x1
x2 <- apply(x1,1, sum)

test_prop_3 <- function (var1,var2){
  frec <- table(var1,var2)[,2:3]
  tot <- apply(frec,1,sum)
  return(list(frec,tot,prop.test(x=c(frec[,2]), n=c(tot), alternative="two.side", conf.level=0.99, correct=TRUE))) 
  # Cogemos la categoría Sí para saber la prop de i en cada población
}
objeto <- c()

for (i in c(55, 56,58:60,64,65)){
  objeto[i] <- list(test_prop_3(Analisis_Poblacion$Poblacion, Analisis_Poblacion[,i]))
  cat ("Tabla y test de proporciones de la variable",names(Analisis_Poblacion)[i], "\n")
  print(objeto[i])
}


```

Tabla A.3. Resultados del test de proporciones para los factores de riesgo para la TB estudiados. Se muestra el nombre del factor de riesgo, el p.valor, si hay diferencia significativa con un nivel de significancia de 0.05 y de 0.01 y qué muestra tiene una proporción mayor de casos con el factor de riesgo correspondiente. Los datos de TB se agregan a nivel global (de CV). Se muestran en negrita los p.valores menores al nivel de significancia 0.05.

| Variable 2 cat                 | p.value      |Sig. al 99%   | Sig. al 95% | Prop. Mayor    |
| -------------------------------|--------------|--------------|-------------|----------------|
| Sexo_H                         |**0.02896 **  |    No        |    Sí       | Sí Genom       |
| Extranjero                     |  0.178       |    No        |    No       |  --            |
| Ingreso_Hospitalario           |**2.381e-09** |    Sí        |    Sí       | Sí Genom       |
| Personal_Sanitario             |  0.2701      |    No        |    No       | --             |
| Alcoholismo                    |**1.391e-09** |    Sí        |    Sí       | Sí Genom       |
| Diabetes                       |  0.175       |    No        |    No       |  --            |
| UDPV                           |**0.02131**   |    No        |    Sí       | Sí Genom       |
| VIH                            |  0.1246      |    No        |    No       | --             |
| Marginado.a                    |**8.324e-05** |    Sí        |    Sí       | Sí Genom       |
| Prisiones                      |**0.0001561** |    Sí        |    Sí       | Sí Genom       |
| Residencia_Ancianos            |  0.1952      |    No        |    No       |  --            |


A nivel global, la proporción de hombres, de ingresos hospitalarios, de casos de alcoholismo, de consumo de drogas por vía intravenosa, personas en riesgo de marginación y personas que han sido ingresadas en prisión es superior en la muestra de casos de TB con información genómica asociada (Tabla A.3). 

### Por provincias

```{r}
# Variable Sexo 

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Sexo)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Sexo)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.sex <- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,1],union1[i,1]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.sex[i] <- (prop.test(x=c(union2[i,1],union1[i,1]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}

```

```{r}

# Variable Extranjero por provincia 
x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Extranjero)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Extranjero)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.ext <- c()
vector.E <- c(0)


for (i in 1:3){
  vector.E[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.ext[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector.E[i])
}

```

```{r}
# Variable Ingreso_Hospitalario por provincia

x1 <- table( Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Ingreso_Hospitalario)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Ingreso_Hospitalario)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.ingr<- c()
vector.Ingr <- c(0)


for (i in 1:3){
  vector.Ingr[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.ingr[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector.Ingr[i])
}

```

```{r}

# Variable Personal_Sanitario por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Personal_Sanitario)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Personal_Sanitario)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.sanit<- c()
vector.Sanit <- c(0)


for (i in 1:3){
  vector.Sanit[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.sanit[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector.Sanit[i])
}
```

```{r}
# Se crea un data.frame con los p.valores de los tests de proporciones
prov.sex <- as.data.frame(prov.sex)

rownames(prov.sex) <- rownames(x1)
prov.val <- prov.sex
prov.val$prov.ext <- prov.ext
prov.val$prov.ingr <- prov.ingr
prov.val$prov.sanit <- prov.sanit

prov.val <- round(prov.val, 3)

# Se arreglan los nombres de las variables
colnames(prov.val)[1] <- "Sexo_H"
colnames(prov.val)[2] <- "Extranjero"
colnames(prov.val)[3] <- "Ingr_Hospital"
colnames(prov.val)[4] <- "Pers_Sanitario"
```

```{r}
# Variable Alcoholismo por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Alcoholismo)[,2:3] #Aquí se corrige el nivel desconocido
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Alcoholismo)[,2:3] #Aquí se corrige el nivel desconocido
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.alc <- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.alc[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```

```{r}

## Variable Diabetes por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Diabetes)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Diabetes)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)


prov.diab <- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.diab[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}

```

```{r}

# Variable UDVP por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$UDVP)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$UDVP)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)


prov.UDVP<- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.UDVP[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
 cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```

```{r}
# Variable VIH por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$VIH)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
(union1 <- cbind(x1, total1))

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$VIH)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
(union2 <- cbind(x2, total2))


prov.VIH<- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.VIH[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
 cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```

```{r}
# Variable Marginado por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Marginado.a)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Marginado.a)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

prov.marg<- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.marg[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```

```{r}
# Variable Prisiones por provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Prisiones)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Prisiones)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)


prov.pris<- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.pris[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```

```{r}
# Ancianos Provincia

x1 <- table(Faltantes_Genom$Provincia_Declaracion, Faltantes_Genom$Residencia_Ancianos)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Provincia_Declaracion, Datos_Genom$Residencia_Ancianos)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)


prov.anc<- c()

vector <- c(0)
for (i in 1:3){
  vector[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  prov.anc[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones provincia", levels(Faltantes_Genom$Provincia_Declaracion)[i])
  print(vector[i])
}
```


```{r}

# Se dividen las variables en 2 tablas para que quepa bien al
# compilar en pdf.

prov.val$prov.alc <- round(prov.alc,3)
prov.val$prov.diab <- round(prov.diab,3)

colnames(prov.val)[5] <- "Alcoholismo"
colnames(prov.val)[6] <- "Diabetes"

prov.val3<- as.data.frame(prov.UDVP)
rownames(prov.val3) <- rownames(x1)
prov.val3$prov.VIH <- prov.VIH
prov.val3$prov.marg <- prov.marg
prov.val3$prov.pris <- prov.pris
prov.val3$prov.anc <- prov.anc


prov.val3 <- round(prov.val3, 3)

colnames(prov.val3)[1] <- "UDPV"
colnames(prov.val3)[2] <- "VIH"
colnames(prov.val3)[3] <- "Margi"
colnames(prov.val3)[4] <- "Prisiones"
colnames(prov.val3)[5] <- "Res_Ancianos"
```


```{r, include=TRUE, echo=FALSE}
kable(prov.val, "latex",
      caption = "\\label{tabla07} p.valores obtenidos del test de proporciones al comparar la proporción de casos de TB para cada uno de los factores de riesgo entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada. Los datos de TB con o sin el factor de riesgo se agregan por provincias." , 
      booktabs = T, label="", align = "c") %>%
  column_spec(2, bold = ifelse(prov.val$Sexo_H > 0.05, FALSE, TRUE)) %>% 
  column_spec(3, bold = ifelse(prov.val$Extranjero > 0.05, FALSE, TRUE)) %>% 
  column_spec(4, bold = ifelse(prov.val$Ingr_Hospital > 0.05, FALSE, TRUE)) %>% 
  column_spec(5, bold = ifelse(prov.val$Pers_Sanitario> 0.05, FALSE, TRUE)) %>% 
  column_spec(6, bold = ifelse(prov.val$Alcoholismo > 0.05, FALSE, TRUE)) %>% 
  column_spec(7, bold = ifelse(prov.val$Diabetes > 0.05, FALSE, TRUE)) %>% 
  kable_styling(full_width = FALSE, position = "center", font_size = 8)
```

```{r, include=TRUE, echo=FALSE}
kable(prov.val3, "latex",
      booktabs = T, label="", align = "c") %>%
  column_spec(2, bold = ifelse(prov.val3$UDPV > 0.05, FALSE, TRUE)) %>% 
  column_spec(3, bold = ifelse(prov.val3$VIH > 0.05, FALSE, TRUE)) %>% 
  column_spec(4, bold = ifelse(prov.val3$Marg> 0.05, FALSE, TRUE)) %>% 
  column_spec(5, bold = ifelse(prov.val3$Prisiones > 0.05, FALSE, TRUE)) %>% 
  column_spec(6, bold = ifelse(prov.val3$Res_Ancianos > 0.05, FALSE, TRUE)) %>% 
  kable_styling(full_width = FALSE, position = "center", font_size = 8)
```

A nivel provincial (Tabla \ref{tabla07}), es posible concluir que la proporción de ingresos_hospitalarios, casos de alcoholismo y personas en riesgo de marginación presenta diferencias significativas al 99% entre ambas muestras. En todos los casos la proporción de afectados es mayor en la muestra de casos de TB con información genómica asociada. Además, Alicante parece ser la provincia que más diferencias acumula  (Tabla \ref{tabla07}). Estos resultados de nuevo apoyan la hipótesis de las características sociales de las dos muestras analizadas son diferentes. 

### Por departamentos de salud

Finalmente, a nivel de departamento de salud, se encuentran diferencias significativas en las proporciones de ingresados en hospital, de casos de alcoholismo, de HIV y en casos en riesgo de marginación (Tabla \ref{tabla09}) . 

Cabe destacar que el prop.test utiliza un chi-cuadrado ymarca que esta aproximación puede no ser adecuada ya que el número de casos esperados para algunos factores en algunos municipios es menor que 5. Se ha utilizado la corrección de Yates, pero aún así la potencia de los test se reduce al desagregar los datos en departamentos de salud. Esto se debe tener en cuenta en la interpretación de los resultados. 

Es difícil afinar en qué factores difieren a nivel de departamento de salud. Sin embargo, los casos de ingresos hospitalarios, de alcoholismo y de personas en riesgo de marginación social sí que difieren tanto a nivel general, provincial y de departamentos de salud entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada. Por tanto, es posible concluir que hay diferencias en la composición social de ambas muestras. 

```{r}
# Variable Sexo por departamentos 

x1 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Sexo)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Sexo)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)


p.val.sex <- c()

vector <- c(0)
for (i in 1:24){
  vector[i] <- list(prop.test(x=c(union2[i,1],union1[i,1]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.sex[i] <- (prop.test(x=c(union2[i,1],union1[i,1]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector[i])
}

```


```{r}

# Variable Extranjero por departamentos 
x1 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Extranjero)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Extranjero)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.ext <- c()
vector.E <- c(0)


for (i in 1:24){
  vector.E[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.ext[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.E[i])
}

```

```{r}
# Variable Ingreso_Hospitalario por departamentos

x1 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Ingreso_Hospitalario)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Ingreso_Hospitalario)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.ingr<- c()
vector.Ingr <- c(0)


for (i in 1:24){
  vector.Ingr[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.ingr[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.Ingr[i])
}
```

```{r}

# Variable Personal_Sanitario por departamentos

x1 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Personal_Sanitario)
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Personal_Sanitario)
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.sanit<- c()
vector.Sanit <- c(0)


for (i in 1:24){
  vector.Sanit[i] <- list(prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.sanit[i] <- (prop.test(x=c(union2[i,2],union1[i,2]), n=c(union2[i,3], union1[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.Sanit[i])
}
```

```{r}
# Se crea un data.frame con los p.valores de 
# los test de proporciones
p.val.sex <- as.data.frame(p.val.sex)

rownames(p.val.sex) <- rownames(x1)
p.valores <- p.val.sex
p.valores$p.val.Ext <- p.val.ext
p.valores$p.val.Ingr <- p.val.ingr
p.valores$p.val.Sanit <- p.val.sanit


# Se redondean los p.valores
p.valores <- round(p.valores, 3)

# Se arreglan los nombres
colnames(p.valores)[1] <- "Sexo_H"
colnames(p.valores)[2] <- "Extranjero"
colnames(p.valores)[3] <- "Ingr_Hospital"
colnames(p.valores)[4] <- "Pers_Sanitario"

```

```{r}
## Variable de 3 categorías por departamento de salud

# Alcoholismo
x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Alcoholismo)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Alcoholismo)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.alc <- c()
vector.a <- c(0)


for (i in 1:24){
  vector.a[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.alc[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.a[i])
}
```


```{r}
# Diabetes
x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Diabetes)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Diabetes)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.diab <- c()
vector.a <- c(0)

for (i in 1:24){
  vector.a[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.diab[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.a[i])
}

```

```{r}

# UDVP

x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$UDVP)[,2:3] # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$UDVP)[,2:3] # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.UDPV <- c()
vector.u <- c(0)


for (i in 1:24){
  vector.u[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.UDPV[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.u[i])
}
```

```{r}
# VIH

x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$VIH)[,2:3]  # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
union1 <- cbind(x1, total1)

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$VIH)[,2:3]  # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
union2 <- cbind(x2, total2)

p.val.VIH <- c()
vector.vih <- c(0)


for (i in 1:24){
  vector.vih[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.VIH[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.vih[i])
}
```

```{r}
# Marginado

x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Marginado.a)[,2:3]  # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
(union1 <- cbind(x1, total1))

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Marginado.a)[,2:3]  # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
(union2 <- cbind(x2, total2))

p.val.marg <- c()
vector.m <- c(0)


for (i in 1:24){
  vector.m[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.marg[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.m[i])
}
```

```{r}
# Prisiones

x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Prisiones)[,2:3]  # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
(union1 <- cbind(x1, total1))

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Prisiones)[,2:3]  # Se corrige nivel "Desconocido"
total2 <- as.matrix(apply(x2,1,sum))
(union2 <- cbind(x2, total2))

p.val.pris <- c()
vector.p <- c(0)


for (i in 1:24){
  vector.p[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.pris[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.p[i])
}
```


```{r}
# Residencia_Ancianos

x1 <- table(Datos_Genom$Dpto_Riesgo, Datos_Genom$Residencia_Ancianos)[,2:3]  # Se corrige nivel "Desconocido"
total1 <- as.matrix(apply(x1,1,sum))
(union1 <- cbind(x1, total1))

x2 <- table(Faltantes_Genom$Dpto_Riesgo, Faltantes_Genom$Residencia_Ancianos)
total2 <- as.matrix(apply(x2,1,sum))
(union2 <- cbind(x2, total2))

p.val.anc <- c()
vector.anc <- c(0)


for (i in 1:24){
  vector.anc[i] <- list(prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))
  p.val.anc[i] <- (prop.test(x=c(union1[i,2],union2[i,2]), n=c(union1[i,3], union2[i,3]), alternative="two.side", conf.level=0.99, correct=TRUE))$p.value
  cat("Test de proporciones Departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
  print(vector.anc[i])
}
```

```{r}

rm(p.valores3)

# Se separan las variables en 2 tablas por motivos de espacio

p.valores$p.val.alc <- round(p.val.alc,3)
p.valores$p.val.diab <- round(p.val.diab,3)

colnames(p.valores)[5] <- "Alcoholismo"
colnames(p.valores)[6] <- "Diabetes"

p.val.UDPV <- as.data.frame(p.val.UDPV)
rownames(p.val.UDPV) <- rownames(x1)

p.valores3 <- p.val.UDPV
p.valores3$p.val.VIH <- p.val.VIH
p.valores3$p.val.marg <- p.val.marg
p.valores3$p.val.pris <- p.val.pris
p.valores3$p.val.anc <- p.val.anc

# Se arreglan los nombres
colnames(p.valores3)[1] <- "UDPV"
colnames(p.valores3)[2] <- "VIH"
colnames(p.valores3)[3] <- "Marg"
colnames(p.valores3)[4] <- "Prisiones"
colnames(p.valores3)[5] <- "Res_Ancianos"

p.valores3 <- round(p.valores3,3)
```
```{r, include=TRUE, echo=FALSE}

kable(p.valores, "latex",
      caption = "\\label{tabla08}  p.valores obtenidos del test de proporciones al comparar la proporción de casos de TB para cada uno de los factores de riesgo entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada. Los datos de TB con o sin el factor de riesgo se agregan por departamento de riesgo. Los casos en los que hay un NaN se deben a que la proporción del factor en ambas submuestras es 0." , 
      booktabs = T, label="", align = "c") %>%
  column_spec(2, bold = ifelse(p.valores$Sexo_H > 0.05, FALSE, TRUE)) %>% 
  column_spec(3, bold = ifelse(p.valores$Extranjero > 0.05, FALSE, TRUE)) %>% 
  column_spec(4, bold = ifelse(p.valores$Ingr_Hospital > 0.05, FALSE, TRUE)) %>% 
  column_spec(5, bold = ifelse(p.valores$Pers_Sanitario <0.05 & !is.na(p.valores$Pers_Sanitario), TRUE, FALSE)) %>%
  column_spec(6, bold = ifelse(p.valores$Alcoholismo > 0.05, FALSE, TRUE)) %>% 
  column_spec(7, bold = ifelse(p.valores$Diabetes > 0.05, FALSE, TRUE)) %>% 
  kable_styling(full_width = FALSE, position = "center", font_size = 8)
```

```{r, include=TRUE, echo=FALSE}
kable(p.valores3, "latex",
      booktabs = T, label="", align = "c") %>%
  column_spec(2, bold = ifelse(p.valores3$UDPV <0.05 & !is.na(p.valores3$UDPV), TRUE, FALSE)) %>% 
  column_spec(3, bold = ifelse(p.valores3$VIH <0.05 & !is.na(p.valores3$VIH), TRUE, FALSE)) %>% 
  column_spec(4, bold = ifelse(p.valores3$Marg <0.05 & !is.na(p.valores3$Marg), TRUE, FALSE)) %>%  
  column_spec(5, bold = ifelse(p.valores3$Prisiones <0.05 & !is.na(p.valores3$Prisiones), TRUE, FALSE)) %>%  
  column_spec(6, bold = ifelse(p.valores3$Res_Ancianos <0.05 & !is.na(p.valores3$Res_Ancianos), TRUE, FALSE)) %>%  
  kable_styling(full_width = FALSE, position = "center", font_size = 8)
```
\clearpage

## Análisis comparativo de la asociación a brote por el método epidemiológico de los casos de TB


```{r, results='hide'}
# Distribución de la asociación a brote

table(Faltantes_Genom$Asociacion_a_Brote)
table(Datos_Genom$Asociacion_a_Brote)

table(Faltantes_Genom$Asociacion_a_Brote)/1372 
table(Datos_Genom$Asociacion_a_Brote)/1632
```

Tabla A.6. Comparación de los porcentajes de casos aislados y pertenecientes a brote en las bases de datos global  entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada. 

| Base de datos   | % Caso aislado  | % Brote      |
|-----------------|-----------------|--------------|
| No Genom.       |     86.22       |    13.78     |
| Sí Genom.       |     86.40       |    13.60     | 


* Nota: El nivel Pendiente del factor se ha omitido, tanto en el factor como en el global de los datos, ya que equivalen al resto de NAs de la base de datos. 

```{r, include=TRUE}
prop.test(x=c(1189,1410), n=c(1379,1632),alternative = "two.sided", conf.level = 0.95 )
```
No hay evidencia suficiente para descartar que las proporciones de casos aislados y brote sean iguales en ambas poblaciones.



```{r}
# Asociación a brote por provincia
Mat_Prov_F <- table(Faltantes_Genom$Asociacion_a_Brote, Faltantes_Genom$Provincia_Declaracion)
Mat_Prov_G <- table(Datos_Genom$Asociacion_a_Brote, Datos_Genom$Provincia_Declaracion)

p.val.prov <- c()
lista.prov <- list()

for ( i in 1:3){
  lista.prov[i]<- list(prop.test(x=c(Mat_Prov_G[1,i],Mat_Prov_F[1,i]), n=c( sum(Mat_Prov_G[c(1,3),i]), sum(Mat_Prov_F[c(1,3),i]))))
  p.val.prov[i] <- prop.test(x=c(Mat_Prov_G[1,i],Mat_Prov_F[1,i]), n=c( sum(Mat_Prov_G[c(1,3),i]), sum(Mat_Prov_F[c(1,3),i])))$p.value
  print(lista.prov[i])
  
}

lista.prov
p.val.prov # Significativo en Valencia

df2<- table(Faltantes_Genom$Asociacion_a_Brote, Faltantes_Genom$Provincia_Declaracion)[c(1,3),]
df1 <- table(Datos_Genom$Asociacion_a_Brote, Datos_Genom$Provincia_Declaracion)[c(1,3),]

Brote_Provincia <- rbind(df1,df2)

# Asociación a brote por departamento de salud

Mat_Genom1 <- table(Datos_Genom$Asociacion_a_Brote, Datos_Genom$Dpto_Riesgo)
Mat_Faltantes1 <- table(Faltantes_Genom$Asociacion_a_Brote, Faltantes_Genom$Dpto_Riesgo)

prop.test(x=c(Mat_Genom1[1,1],Mat_Faltantes1[1,1]), n=c( sum(Mat_Genom1[c(1,3),1]), sum(Mat_Faltantes1[c(1,3),1])))

p.val.R <- c()
lista <- c()
for ( i in 1:24){
  lista[i]<- list(prop.test(x=c(Mat_Genom1[1,i],Mat_Faltantes1[1,i]), n=c( sum(Mat_Genom1[c(1,3),i]), sum(Mat_Faltantes1[c(1,3),i]))))
  p.val.R[i] <- prop.test(x=c(Mat_Genom1[1,i],Mat_Faltantes1[1,i]), n=c( sum(Mat_Genom1[c(1,3),i]), sum(Mat_Faltantes1[c(1,3),i])))$p.value
cat("Test de proporciones Departamento", levels(Faltantes_Genom$Dpto_Riesgo)[i])
print(lista[i])
  
}


p.val.R # Ninguno significativo 
```


```{r, include=TRUE}
p.val.prov
Provincia <- c("Alicante", "Castellón", "Valencia")
AB1 <- data.frame("Provincia"=Provincia, "P_Valores"=p.val.prov)

p.val.R 
AB2 <- data.frame("Dpto_Riesgo"=colnames(Mat_Genom1), "P_Valores"=p.val.R)
```

```{r, echo=FALSE, include=TRUE}
AB1 %>%
  kable(booktabs = TRUE,format = "latex",
        caption = "\\label{tabla04}  p.valores obtenidos en el test de proporciones al comparar la proporción de casos de TB asociados a brote          entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada a nivel         de provincia")%>%
  column_spec(1, width = "3cm") %>%
  kable_styling(
    latex_options = c("striped", "condensed","hold_position"),
    position = "center",
    full_width = FALSE)
```

```{r, echo=FALSE, include=TRUE}
AB2 %>%
  kable(booktabs = TRUE,
        format = "latex",
          caption = "\\label{tabla05}  P.valores obtenidos en el test de proporciones al comparar la proporción de casos de TB asociados a brote           entre la muestra de casos de TB con información genómica asociada y la muestra de casos de TB sin información genómica asociada a               nivel de departamentos de salud") %>%
  column_spec(1, width = "8cm") %>%
  kable_styling(
    latex_options = c("striped", "condensed", "hold_position"),
    position = "center",
    full_width = FALSE)
```
