==> scripts/06_Analisis_Complementarios.Rmd <==
---
title: "06_Analisis_Complementario"
author: "Paula Sinisterra Sebastián"
date: "2023-04-14"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, include = FALSE)
```


```{r}
## Carga de librerías

library(readr) # Importar documentos
library(readxl) # Importar documentos
library(DCluster) # Modelo Poisson-Gamma
library(R2WinBUGS) # Modelos BYM
library(gridExtra) # unir mapas 
library(sp) # Elaboración de mapas
library(dplyr)
library(kableExtra) # Elaboración de tablas
library(caret) # Dividir un grupo de datos
library(pROC) # Curvas ROC
library(ResourceSelection)  #hoslem.test
library(RColorBrewer) # Colorear mapas
```


```{r}
# Carga de datos

Obs_Muni <- read_excel("../results/Obs_Muni.xlsx")
Obs_Muni_Genom <- read_excel("../results/Obs_Muni_Genom.xlsx")

Trans_Muni <- read_excel("../results/Trans_Muni.xlsx")
```

```{r}
# Carga de cartografía
load("../maps/CartoCV.Rdata")

Carto_Muni <- carto_muni
Carto_Muni=Carto_Muni[order(Carto_Muni@data$INE_MUN),] 
 as.data.frame(Carto_Muni)
 
 # Estructura de vecindad
Carto_Muni.nb <- poly2nb(Carto_Muni)
Carto_Muni$ORDEN <- 1:542


# Unión "islas"
Carto_Muni.nb[[477]] <- as.integer(c(277,312))
Carto_Muni.nb[[312]] <- as.integer(c(317,382, 477, 517, 523, 538))

Carto_Muni.nb[[364]] <- as.integer(c(317,363,528))
Carto_Muni.nb[[317]] <- as.integer(c(312,364,517,523))

# Relación de vecindad
Vecinos <- nb2WB(Carto_Muni.nb)
```



# Relación entre la variable respuesta (Transmisión) y los factores de la base de datos

```{r}
# Base de datos arreglada
Datos_Codigos <- read_csv("../data/Datos_Codigos.csv")
Datos_Codigos <- as.data.frame(Datos_Codigos[,c(2:75)])

# data.frame con los datos genómicos
Datos_Genom <- Datos_Codigos[ !is.na(Datos_Codigos$ID_Genom),]

Datos_Genom$Ber <- Datos_Genom$Cluster_5snps
Datos_Genom$Ber[Datos_Genom$Ber=="unique"] <- "NO"
Datos_Genom$Ber[Datos_Genom$Ber!="NO"] <- "SI"

 # No hay NAs para Municipio_Residencia
 which(is.na(Datos_Genom$Municipio_Residencia))

 for (i in 2:75){
  Datos_Genom[,i]<- as.factor(Datos_Genom[,i])
}

```

Hay 75 variables en la base de datos, pero algunas están repetidas o tienen demasiados NAS (>400). Por eso, para analizar las relaciones con la variable respuesta se utiizan solo las siguientes:

* Transmisión(Ber) como variable respuesta

* Year_declaration
* Dto_Declaration
* S_Declaration
* Proc_Declaration
* No_Edad
* Sexo
* Extranjero
* Asociacion_a_brote
* Provincia_Declaracion
* No_Convivientes (muchos NAs)
* Clasificacion_Caso
* Tipo_de_Caso
* Tincion
* Cultivo
* Rx_Torax
* Tipo_Germen
* Localizacion
* presencia_Síntomas
* Situacion_Evolutiva
* ResI
* ResP
* ResE
* ResR
* ResS
* Ingreso_Hospitalario
* Alcoholismo
* Diabetes
* Neoplasia
* UDVP  
* VIH   
* Marginado.a
* Inmigrante
* Transeunte   
* Personal_Sanitario               
* Prisiones     
* Residencia_Ancianos    
* TOD   
* Evolucion  
* Estudio_Contactos
* No_Contactos_Infectados    

Al analizar los NAs se observa que hay muchas variables que contienen diferente número de valores faltantes, se debe tener esto presente en el análisis.

```{r}
colSums(is.na(Datos_Genom))
```
## 1. Relación individual la variable respuesta (Transmisión) y los factores de la base de datos

Se utiliza el **test chi-cuadrado** para concluir si dos variables categóricas son independientes entre si (Ho). En el caso de rechazar la hipótesis nula, se asumirá que hay algún tipo de dependencia entre las variables. En este caso la variable de interés es Transmisión (Factor Sí/No) y se quiere saber si es independiente a cada una de las covariables, o no. *Nota*: no se utiliza el Test de Fisher porque las marginales no han sido fijadas por el experimento y tenemos tamaños muestrales relativamente grandes. 

```{r}
# Se seleccionan las variables con las que se realiza el análisis
Datos_Genom_RV <- Datos_Genom[, c(4,6,8,9,11,13,14,15,18,31, 32,35:40,43:50, 55:65,67,68,69,71,75 )]
```

```{r}

# Test chi-cuadrado para variables de 2 categorías
p.val2 <- c()

for (i in c(6,7,25,34,37,39)){
  
  Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV[,i]),] # Se eliminan los NAs
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i]))
  cat("Resultado chisq test", names(Datos_Genom_RV)[i])
  print(chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])))
  p.val2[i] <- chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i]))$p.value
}

p.val2 <- na.omit(p.val2)

# Test chi-cuadrado para variables de 3 categorías
# las categorías son Sí, No y Desconocido
# Así que los desconocidos se tratan como NAs

p.val3 <- c()
for (i in c(18,20:24,26:33,35,36)){
  
  Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV[,i]),]
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,2:3]) # Se corrige el nivel "Desconocido"
  cat("Resultado chisq test", names(Datos_Genom_RV)[i])
  print(chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,2:3])) # Se corrige el nivel "Desconocido"
  p.val3[i] <- chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,2:3])$p.value
  
}

p.val3 <- na.omit(p.val3)

# Test chi-cuadrado para variables de 4 categorías

p.val4 <- c()
for (i in c(10,13)){
  
  Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV[,i]),] # Se eliminan los NAs
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,c(1,3)]) # Se corrigen los niveles equivalentes a NAs
  cat("Resultado chisq test", names(Datos_Genom_RV)[i])
  print(chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,c(1,3)]))
  p.val4[i] <- chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])[,c(1,3)])$p.value
  
}

p.val4 <- na.omit(p.val4)


# Test chi-cuadrado para variables de 5 categorías

p.val5 <- c()
for (i in c(1,9)){
  Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV[,i]),] # Se eliminan los NAs
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i]))
  cat("Resultado chisq test", names(Datos_Genom_RV)[i])
  print(chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i])))
  p.val5[i] <- chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio[,i]))$p.value
}

p.val5 <- na.omit(p.val5)

# Casos en que hay que poner atención especial

# Asociación_a_Brote
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Asociacion_a_Brote),]
Datos_Genom_Limpio <- Datos_Genom_RV[!Datos_Genom_RV$Asociacion_a_Brote=="Pendiente",]
Datos_Genom_Limpio <- droplevels(Datos_Genom_Limpio)
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Asociacion_a_Brote))  # < 2.2e-16
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Asociacion_a_Brote))$expected
  
#Tipo de Caso
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Tipo_de_Caso),]
Datos_Genom_Limpio$Tipo_de_Caso <- droplevels(Datos_Genom_Limpio$Tipo_de_Caso)
 chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Tipo_de_Caso)) #0.0002423
 
 
# Tinción
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Tincion),]
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Tincion)[,c(1,4)]) #0.02815
  
# Rx_Torax
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Rx_Torax),]
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Rx_Torax)[,c(2,4)]) #2.091e-05 

   # Tipo_Germen 
Tipo_Germen <- Datos_Genom_RV$Tipo_Germen
Tipo_Germen.sn=vector() 
Tipo_Germen.sn[Tipo_Germen=="M. Tuberculosis"]=0 
Tipo_Germen.sn[Tipo_Germen=="M. Bovis" | Tipo_Germen=="M. Africanum" | Tipo_Germen=="Otras" | Tipo_Germen=="Sin tipificar"]=1 
Tipo_Germen.sn=factor(Tipo_Germen.sn,labels=c("M.tuberculosis","Otras"))
Datos_Genom_RV$Tipo_Germen.sn=Tipo_Germen.sn

Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Tipo_Germen.sn),]
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Tipo_Germen.sn)) #  0.3853

  
#Localizacion
Localizacion <- Datos_Genom_RV$Localizacion
Localizacion.sn=vector() 
Localizacion.sn[Localizacion=="Pulmonar"]=0 
Localizacion.sn[Localizacion=="Digestiva" | Localizacion=="Diseminada" | Localizacion=="Genitourinaria" | Localizacion=="Laringea"|Localizacion=="Linfatica"|Localizacion=="Linfática"|Localizacion=="Linfatica (Extratoracica)"| Localizacion=="Linfática (Extratorácica)" | Localizacion=="Meningitis Tuberculosa"|Localizacion=="Miliar"| Localizacion=="Osteoarticular"|Localizacion=="Otras"|Localizacion=="Pleural"|Localizacion=="Sistema Nervioso Cental"]=1 

Localizacion.sn=factor(Localizacion.sn,labels=c("Pulmonar","No_Pulmonar")) 
Datos_Genom_RV$Localizacion.sn=Localizacion.sn 

levels(Datos_Genom_RV$Localizacion.sn)
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Localizacion.sn),]
  chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Localizacion.sn)) #2.018e-09


#Presencia de Síntomas
Datos_Genom_RV$Presencia_de_sintomas[Datos_Genom_RV$Presencia_de_sintomas=="Sí"] <- "Si"
Datos_Genom_RV$Presencia_de_sintomas <- droplevels(Datos_Genom_RV$Presencia_de_sintomas)
Datos_Genom_Limpio <- Datos_Genom_RV[!is.na(Datos_Genom_RV$Presencia_de_sintomas),]
chisq.test(table(Datos_Genom_Limpio$Ber, Datos_Genom_Limpio$Presencia_de_sintomas)[,2:3]) #0.9342

```

```{r}

# Se elabora un data.frame con los resultados

p.val <- c(p.val2,p.val3, p.val4, p.val5, 2.2e-16, 0.0002423, 0.02815, 2.091e-05 , 0.3853, 2.018e-09, 0.9342)

Nombres1 <- as.character()
for (i in c(6,7,25,34,37,39)){
  Nombres1[i] <-  names(Datos_Genom_RV)[i]
}

Nombres1 <- na.omit(Nombres1)

Nombres2 <- as.character()
for (i in c(18,20:24,26:33,35,36)){
  Nombres2[i] <- names(Datos_Genom_RV)[i]
}

Nombres2 <- na.omit(Nombres2)

Nombres3 <- as.character()
for ( i in c(10,13)){
  Nombres3[i]<- names(Datos_Genom_RV)[i]
}
Nombres3 <- na.omit(Nombres3)

Nombres4 <- as.character()
for ( i in c(1,9)){
  Nombres4[i]<- names(Datos_Genom_RV)[i]
}
Nombres4 <- na.omit(Nombres4)

Nombres5 <- as.character()
for ( i in c(8,11,12,14,15,16,17)){
  Nombres5[i]<- names(Datos_Genom_RV)[i]
}
Nombres5 <- na.omit(Nombres5)

Nombres <- c(Nombres1,Nombres2, Nombres3,Nombres4, Nombres5)

Result_Chi<- data.frame("Nombres"=Nombres, "P.val"=p.val)
```

```{r, include=TRUE}
kable(Result_Chi)
```

Por tanto, los resultados de la Tabla indican que las variables **Extranjero**, **TOD**, **Alcoholismo**, **Inmigrante**, **Prisiones**,**Rx_Torax**,**Tipo_de_Caso**, **Localización*** y **Asociación a Brote** parecen no ser independientes de la variable Transmisión (**Ber**). Además, las variables **Sexo**, **Ingreso_Hospitalario**, **Situacion_Evolutiva**, **Marginado.a** y **Tincion**tienen un p.valor inferior a 0.05, así que al 95% de significatividad también se puede rechazar que sean independientes de la variable Transmisión. 


## 2. Regresión Logística

Se realiza una regresión logística para analizar la importancia de las variables en conjunto y  poder estudiar la calidad predictiva del modelo. 

Para poder llevar a cabo los modelos de regresión, tenemos que tener variables SIN NAs. Por este motivo, en primer lugar, se prueba **un modelo solo con las variables que no tienenningún NA**:

### Primer estrategia: utilizar las variables sin datos faltantes

**Modelo 1**

$$Logit(Transmision) \sim year.declaracion + S.Declaracion + Proc.Declaracion + No.Edad + Sexo  + Cultivo $$

```{r}
# Se convierte la variable Ber a dummy 
dummy_survived <- model.matrix (~Datos_Genom$Ber)
Datos_Genom$Ber  <- dummy_survived[,2]
Datos_Genom$Ber   <- Datos_Genom$Ber 
str(Datos_Genom$Ber)
```


```{r}
#Datos_Genom

Datos_Genom$year_declaracion <- as.numeric(Datos_Genom$year_declaracion)
Datos_Genom$S_Declaracion <- as.numeric(Datos_Genom$S_Declaracion)
Datos_Genom$No_Edad <- as.numeric(Datos_Genom$No_Edad)
```



```{r}
### Se separan los grupos test y train
set.seed(123)

# Se crean los índices de las observaciones de entrenamiento
train <- createDataPartition(y = Datos_Genom$Ber, p = 0.8, list = FALSE, times = 1)
train1 <- Datos_Genom[train, ]
test1 <- Datos_Genom[-train, ]

#Para verificar que los subgrupos se parecen al grupo mayor:
prop.table(table(Datos_Genom$Ber))
prop.table(table(train1$Ber))
prop.table(table(test1$Ber))
```


```{r}
# Modelo de regresión logística

Reg_Model1 <- glm(Ber ~ year_declaracion + S_Declaracion + Proc_Declaracion + No_Edad + Sexo  + Cultivo , data=Datos_Genom, family = "binomial"(link = logit))

summary(Reg_Model1)
step(Reg_Model1)


# Modelo obtenido del Step
Reg_Model1_Step <- glm(Ber ~  year_declaracion + No_Edad + Sexo  , data=Datos_Genom, family = "binomial"(link = logit))
summary(Reg_Model1_Step)

# Modelo utilizando solo el grupo train
Reg_Model1_Step_T <- glm(Ber ~  year_declaracion + No_Edad + Sexo  , data=train1, family = "binomial"(link = logit))
summary(Reg_Model1_Step_T)

```



```{r, eval=FALSE}
# Curvas ROC

# Predicciones para el grupo test
test.prob <- predict(Reg_Model1_Step_T,test1,type="response")

## Matriz de confusión
ajuste.train.pred <- rep("0_Surv",177)
ajuste.train.pred[test.prob > 0.5] <- "1_Surv"
table(test1$Ber,ajuste.train.pred)

## curva ROC
pred.values <- predict(Reg_Model1_Step_T,test1, type="response")
curv_roc_Modelo1 <- roc(test1$Ber,pred.values)
plot(curv_roc_Modelo1,xlim=c(1,0))

# Resultado curva ROC
curv_roc_Modelo1$auc
```

```{r}
# Test Hosmer-Lemeshow
hoslem.test(Datos_Genom$Ber , fitted(Reg_Model1_Step)) # es válido
```

```{r}
# Cross-validation
cv.Modelo1 <- cv.glm(Datos_Genom,Reg_Model1_Step)
cv.Modelo1$delta
```

Aunque los resultados de test de Hosmer-Lemeshow indican que no se puede rechazar que el ajuste del modelo sea bueno, la calidad de la predicción de este modelo no es mucho mejor que si se hiciera el reparto por azar. Por este motivo, se recurre a otra estrategia

### Segunda Estrategia: eliminar individuos para utilizar más variables

Como segunda opción para buscar un modelo que pueda predecir mejor, nos planteamos añadir variables, pero se debe quitar filas (NAs) para ello. Se pretende crear un modelo con todas las variables que resultan interesantes (nombradas al principio del documento), por eso, se eliminan primero todas las líneas que contienen NAs para alguna de ellas (nos quedamos con 877 datos de los 1651 iniciales) y se crea el modelo (Reg_Model2).

Sobre este modelo, se aplica la función *Step* para seleccionar las variables que más importancia tengan en el ajuste. Las variables seleccionadas después de aplicar la función Step son:


$$ Logit(Transmision) \sim Localizacion.sn+ Situacion.Evolutiva + year.declaracion + Prisiones + Estudio.Contactos + No.Edad + Sexo +$$
$$ Tipo.de.Caso + Asociacion.a.Brote + Extranjero $$

Al observar el resumen del modelo, las variables *Localización.sn* y *Situacion.Evolutiva* no parecen significativas. Así, que se simplifica el modelo eliminando esta variable:

**Modelo 2**

$$ Logit(Transmision) \sim year.declaracion + Prisiones + Estudio.Contactos + No.Edad + Sexo + $$
$$  Tipo.de.Caso + Asociacion.a.Brote + Extranjero $$
```{r}
Datos_Genom_RV$year_declaracion <- as.numeric(Datos_Genom_RV$year_declaracion)
Datos_Genom_RV$S_Declaracion <- as.numeric(Datos_Genom_RV$S_Declaracion)
Datos_Genom_RV$No_Edad <- as.numeric(Datos_Genom_RV$No_Edad)
Datos_Genom_RV <- na.omit(Datos_Genom_RV)

```


```{r, include= TRUE, results='hide'}
Reg_Model2 <- glm(Ber ~ year_declaracion + Dpto_Declaracion + S_Declaracion + Proc_Declaracion + No_Edad + Sexo  + Extranjero + Asociacion_a_Brote + Provincia_Declaracion + Clasificacion_Caso + Tipo_de_Caso + Tincion + Rx_Torax  + Tipo_Germen.sn + Localizacion.sn + Presencia_de_sintomas + Situacion_Evolutiva +   resH.INH +  resZ.PZA+  resE.EMB + resR.RIF + resS.SM + Ingreso_Hospitalario + Alcoholismo + Diabetes + Neoplasia + UDVP + VIH + Marginado.a + Inmigrante +Transeunte  +Personal_Sanitario +Prisiones +Residencia_Ancianos +TOD  +Evolucion+ Estudio_Contactos + No_Contactos_Infectados , data=Datos_Genom_RV, family = "binomial"(link = logit))

step(Reg_Model2)

Reg_Model2 <- glm(Ber ~  Localizacion.sn + Situacion_Evolutiva + year_declaracion + Prisiones + Estudio_Contactos + No_Edad + Sexo + Tipo_de_Caso + Asociacion_a_Brote + Extranjero, data=Datos_Genom_RV, family = "binomial"(link = logit))
summary(Reg_Model2)
```

```{r, include=TRUE}
Reg_Model2 <- glm(Ber ~  year_declaracion + Prisiones + Estudio_Contactos + No_Edad + Sexo + Tipo_de_Caso + Asociacion_a_Brote + Extranjero, data=Datos_Genom_RV, family = "binomial"(link = logit))
summary(Reg_Model2)
```

```{r}
# Test 
hoslem.test(Datos_Genom_RV$Ber , fitted(Reg_Model2), g=10)

# Cross-validation
cv.Modelo2 <- cv.glm(Datos_Genom_RV,Reg_Model2)
cv.Modelo2$delta
cv.Modelo1$delta
```

```{r}
# Curva ROC

# Se crean los índices de las observaciones de entrenamiento
set.seed(233)
traind <- createDataPartition(y = Datos_Genom_RV$Ber, p = 0.8, list = FALSE, times = 1)
train2 <- Datos_Genom_RV[traind, ]
test2 <- Datos_Genom_RV[-traind, ]

#Para verificar que los subgrupos se parecen al grupo mayor:
prop.table(table(Datos_Genom_RV$Ber))
prop.table(table(train2$Ber))
prop.table(table(test2$Ber))


Reg_Model2_T <- glm(Ber ~  year_declaracion + Prisiones + Estudio_Contactos + No_Edad + Sexo + Tipo_de_Caso + Asociacion_a_Brote + Extranjero, data=train2, family = "binomial"(link = logit))
summary(Reg_Model2_T)


# Curvas ROC
test.prob <- predict(Reg_Model2_T,test2,type="response")

## Matriz de confusión
ajuste.train.pred <- rep("0_Surv",174)
ajuste.train.pred[test.prob > 0.5] <- "1_Surv"
table(test2$Ber,ajuste.train.pred)

## roc
pred.values <- predict(Reg_Model2_T,test2, type="response")
curv_roc_Modelo2 <- roc(test2$Ber,pred.values)

plot(curv_roc_Modelo2,xlim=c(1,0))

curv_roc_Modelo2$auc
```


Con la base de datos con los casos de TB que tienen información genómica original (que contiene 1651 datos), se eliminan solo los NAs que corresponden a las variables elegidas (51 NAs --> 1600 datos ) y se realiza de nuevo el modelo:

**Modelo 3**

$$ Logit(Transmision) \sim No_Edad + Prisiones +Sexo + Extranjero + Tipo_de_Caso + Asociacion_a_Brote $$

```{r}
Datos_Genom3 <- Datos_Genom[,c("year_declaracion", "Prisiones", "Estudio_Contactos", "No_Edad", "Sexo", "Tipo_de_Caso", "Asociacion_a_Brote", "Extranjero", "Ber")]
Datos_Genom3 <- na.omit(Datos_Genom3)
Datos_Genom3 <- Datos_Genom3[!Datos_Genom3$Prisiones=="Desc",]

# Se separan grupos train y test
set.seed(233)
traind <- createDataPartition(y = Datos_Genom3$Ber, p = 0.8, list = FALSE, times = 1)
train3 <- Datos_Genom3[traind, ]
test3 <- Datos_Genom3[-traind, ]

#Para verificar que los subgrupos se parecen al grupo mayor:
prop.table(table(Datos_Genom3$Ber))
prop.table(table(train3$Ber))
prop.table(table(test3$Ber))

# Modelo 3
Reg_Model3 <- glm(Ber ~  year_declaracion + Prisiones + Estudio_Contactos + No_Edad + Sexo + Tipo_de_Caso + Asociacion_a_Brote + Extranjero, data=Datos_Genom3, family = "binomial"(link = logit))
summary(Reg_Model3)

# Curvas ROC

# Predicciones para el grupo test
test.prob <- predict(Reg_Model3,Datos_Genom3,type="response")

## Matriz de confusión
ajuste.train.pred <- rep("0_Surv",877)
ajuste.train.pred[test.prob > 0.5] <- "1_Surv"
table(Datos_Genom3$Ber,ajuste.train.pred)

## Curva ROC
pred.values <- predict(Reg_Model3,Datos_Genom3, type="response")
curv_roc_Modelo3 <- roc(Datos_Genom3$Ber,pred.values)
plot(curv_roc_Modelo3,xlim=c(1,0))
curv_roc_Modelo3$auc


# Modelo en el grupo train/test
Reg_Model3_S <- glm(Ber ~  year_declaracion + Estudio_Contactos+ No_Edad + Prisiones +Sexo +Extranjero+ Tipo_de_Caso + Asociacion_a_Brote , data=train3, family = "binomial"(link = logit))
summary(Reg_Model3_S)


# Curvas ROC
test.prob <- predict(Reg_Model3_S,test3,type="response")

## Matriz de confusión
ajuste.train.pred <- rep("0_Surv",174)
ajuste.train.pred[test.prob > 0.5] <- "1_Surv"
table(test3$Ber,ajuste.train.pred)

## roc
pred.values <- predict(Reg_Model3_S,test3, type="response")
curv_roc_Modelo3_S <- roc(test3$Ber,pred.values)
plot(curv_roc_Modelo3_S,xlim=c(1,0))
curv_roc_Modelo3_S$auc



```


```{r}
cv.Modelo3 <- cv.glm(Datos_Genom3,Reg_Model3)
cv.Modelo3.1 <- cv.glm(train3,Reg_Model3_S)

cv.Modelo1$delta
cv.Modelo2$delta
cv.Modelo3$delta
cv.Modelo3.1$delta
```

El Modelo 3, que contiene las variables seleccionadas y solo elimina los NAs para esas variables, resulta válido en términos de ajuste y la calidad en términos de predicción. Separando en grupo test y train, se obtiene un AUC de 0.75, es decir, mejor que en el caso del Modelo 1 y el Modelo 2. Sin embargo, dado que para llegar a este modelo se deben eliminar 51 datos y que la curva ROC revela que el modelo no consigue clasificar la transmisión bien a través de los otros factores en bastantes ocasiones. Por este motivo, se rechaza la idea de construir un modelo BYM de imputación para predecir la transmisión en los casos en que no se tiene dicha información.
